Refactor: create new function getReputationForTopDomains
