+++ amplify/backend/function/getReputationForTopDomains/custom-policies.json
@@ -1,0 +1,6 @@
+[
+  {
+    "Action": [],
+    "Resource": []
+  }
+]
