+++ amplify/backend/function/getReputationForTopDomains/function-parameters.json
@@ -1,0 +1,3 @@
+{
+  "lambdaLayers": []
+}
