+++ amplify/backend/function/getReputationForTopDomains/src/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "getuserreputationfortopdomains",
+  "name": "getreputationfortopdomains",
   "version": "2.0.0",
   "description": "Lambda function generated by Amplify",
   "main": "index.js",
