+++ amplify/team-provider-info.json
@@ -18,7 +18,8 @@
         "createUniqueColony": {},
         "createUniqueDomain": {},
         "getUserReputation": {},
-        "getUserReputationForTopDomains": {}
+        "getUserReputationForTopDomains": {},
+        "getReputationForTopDomains": {}
       }
     }
   }
